- WebSecurityConfig: bestellingen en bestelregels verwijderd uit geblockte urls voor klanten - BestellingController & BestelregelController kunnen alle mappings nu afhandelen - domein/Bestelregel heeft nu als id een long ipv een int - domein/klant heeft nu een equals methode - bestelregels.html heeft nu ook een link naar de muteerpagina - bestelregelMuteer.html toegevoegd

+++ src/main/java/springframework/secureWeb/controller/BestellingController.java
@@ -1,6 +1,6 @@
 package springframework.secureWeb.controller;
 
-import java.math.BigDecimal;
+import java.security.Principal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -11,16 +11,21 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.SessionAttributes;
 
+import springframework.secureWeb.data.AccountRepository;
 import springframework.secureWeb.data.ArtikelRepository;
 import springframework.secureWeb.data.BestellingRepository;
 import springframework.secureWeb.data.BestelregelRepository;
 import springframework.secureWeb.data.KlantRepository;
+import springframework.secureWeb.domein.Account;
 import springframework.secureWeb.domein.Artikel;
 import springframework.secureWeb.domein.Bestelling;
 import springframework.secureWeb.domein.Bestelregel;
+import springframework.secureWeb.domein.Klant;
 
 @Controller
+@SessionAttributes({ "bestellingId" })
 public class BestellingController {
 
 	@SuppressWarnings("unused")
@@ -28,67 +33,81 @@ public class BestellingController {
 	private final KlantRepository klantRepo;
 	private final BestelregelRepository bestelregelRepo;
 	private final ArtikelRepository artikelRepo;
+	private final AccountRepository accountRepo;
 
 	@Autowired
 	public BestellingController(BestellingRepository bestellingRepo, KlantRepository klantRepo,
-			BestelregelRepository bestelregelRepo, ArtikelRepository artikelRepo) {
+			BestelregelRepository bestelregelRepo, ArtikelRepository artikelRepo, AccountRepository accountRepo) {
 		this.bestellingRepo = bestellingRepo;
 		this.klantRepo = klantRepo;
 		this.bestelregelRepo = bestelregelRepo;
 		this.artikelRepo = artikelRepo;
+		this.accountRepo = accountRepo;
 	}
 
 	@GetMapping("/bestelling")
-	public String Bestellingen(Model model) {
-		List<Bestelling> bestellingList = new ArrayList<>();
-		bestellingRepo.findAll().forEach(i -> bestellingList.add(i));
+	public String Bestellingen(Model model, Principal principal) {
+		List<Bestelling> bestellingList;
+		if (!ingelogdeAccountIsKlant(principal)) {
+			bestellingList = new ArrayList<>();
+			bestellingRepo.findAll().forEach(i -> bestellingList.add(i));
+
+			model.addAttribute("bestellingId");
+		} else {
+			Klant klant = klantRepo.findKlantByAccount(accountRepo.findByuserNaam(principal.getName()));
+			bestellingList = bestellingRepo.findBestellingByKlant(klant.getId());
+		}
 
 		model.addAttribute("bestellingen", bestellingList);
-		model.addAttribute("spatie", " ");
 
 		return "bestellingen";
 	}
 
-	@GetMapping("/bestelling/{klantId}")
-	public String Bestelregels(Model model, @PathVariable("klantId") String klantId) {
-		try {
-			long klantIdLong = Long.valueOf(klantId);
-			List<Bestelling> bestellingList = bestellingRepo.findBestellingByKlant(klantIdLong);
-
-			model.addAttribute("bestellingen", bestellingList);
-			model.addAttribute("spatie", " ");
-
-			return "bestellingen";
-		} catch (NumberFormatException e) {
-			return "redirect:/main";
-		}
-	}
-
 	@RequestMapping("/bestelling/nieuw")
 	public String nieuweBestelling() {
-		Bestelling bestelling = new Bestelling();
-		// bestelling.setKlant(klantRepo.findById(206L).get());
-		bestelling.setPrijs(new BigDecimal("0"));
-		Bestelling savedBestelling = bestellingRepo.save(bestelling);
-		long bestellingIdLong = savedBestelling.getId();
-		return "redirect:/bestelregel/nieuw/" + bestellingIdLong;
+		return "redirect:/bestelregel/nieuw/";
 	}
 
-	@PostMapping("/bestelling/verwijder/{bestellingId}")
-	public String bestellingVerwijder(@PathVariable("bestellingId") String bestellingId) {
-		long bestellingIdLong = Long.valueOf(bestellingId);
+	/*
+	 * @RequestMapping("/mijnbestelling/nieuw") public String nieuweMijnBestelling()
+	 * { return "redirect:/bestelregel/nieuw/"; }
+	 */
 
-		List<Bestelregel> bestelregelList = bestelregelRepo.findBestelregelByBestelling(bestellingIdLong);
-		for (Bestelregel bestelregel : bestelregelList) {
-			wijzigArtikelVoorraad(bestelregel.getArtikel(), bestelregel.getAantal());
+	@PostMapping("/bestelling/verwijder/{bestellingId}")
+	public String bestellingVerwijder(@PathVariable("bestellingId") String bestellingId, Principal principal) {
+		long bestellingIdLong = Long.valueOf(bestellingId);
+		Bestelling bestelling = bestellingRepo.findById(bestellingIdLong).get();
+
+		if (heeftToegangTotBestelling(principal, bestelling)) {
+			List<Bestelregel> bestelregelList = bestelregelRepo.findBestelregelByBestelling(bestellingIdLong);
+			for (Bestelregel bestelregel : bestelregelList) {
+				wijzigArtikelVoorraad(bestelregel.getArtikel(), bestelregel.getAantal());
+			}
+			bestellingRepo.deleteById(bestellingIdLong);
+			return "redirect:/bestelling";
 		}
-		bestellingRepo.deleteById(bestellingIdLong);
-		return "redirect:/bestelling";
+		return "redirect:/main";
 	}
 
 	private void wijzigArtikelVoorraad(Artikel artikel, int aantal) {
 		artikel.setVoorraad(artikel.getVoorraad() + aantal);
 		artikelRepo.save(artikel);
+	}
+
+	private boolean ingelogdeAccountIsKlant(Principal principal) {
+		return accountRepo.findByuserNaam(principal.getName()).getRol().equals(Account.Rol.klant);
+	}
+
+	private boolean heeftToegangTotBestelling(Principal principal, Bestelling bestelling) {
+		// als ingelogde account een klant is, dan is toegang afhankelijk of de
+		// bestelling ook op diens naam staat
+		if (accountRepo.findByuserNaam(principal.getName()).getRol().equals(Account.Rol.klant)) {
+			Klant ingelogdeKlant = klantRepo.findKlantByAccount(accountRepo.findByuserNaam(principal.getName()));
+			Klant klantOpWieBestellingStaat = bestelling.getKlant();
+			return ingelogdeKlant.equals(klantOpWieBestellingStaat);
+		} else
+			return true;
+
 	}
 
 }
